move code mirror helper back
