+++ packages/insomnia/src/ui/routes/design.tsx
@@ -33,16 +33,6 @@ interface LintMessage extends Notice {
   range: IRuleResult['range'];
 }
 
-CodeMirror.registerHelper('lint', 'openapi', async function(text: string) {
-  const isLintError = (result: IRuleResult) => result.severity === 0;
-
-  const run = await window.main.spectralRun(text);
-  return run.filter(isLintError).map(result => ({
-    from: CodeMirror.Pos(result.range.start.line, result.range.start.character),
-    to: CodeMirror.Pos(result.range.end.line, result.range.end.character),
-    message: result.message,
-  }));
-});
 const RenderEditor: FC<{ editor: RefObject<CodeEditorHandle> }> = ({ editor }) => {
   const activeApiSpec = useSelector(selectActiveApiSpec);
   const [lintMessages, setLintMessages] = useState<LintMessage[]>([]);
